pyproject.toml
[tool.poetry]
name = "link-archive"
version = "0.1.0"
description = ""
authors = ["Tian Xu <tianxu2009@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
